+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## v1.1.0 (2019-01-30)
+
+#### Features
+
+* Upgrade Flurry Android SDK version to 11.5.0
+* Upgrade Flurry iOS SDK version to 9.3.0
+* Use IDictionary instead of Dictionary for API parameters
+
+#### Bug Fixes
+
+* Make iOS log level consistent with Android
+
 # v1.0.0 (2018-12-06)
 
 #### Features